+++ tests/macros.rs
@@ -8,7 +8,7 @@ use bitvec::prelude::*;
 
 #[test]
 fn bits() {
-	let bits: &BitSlice<Local, Word> = bits![1, 0, 1, 1, 0, 1, 0, 0, 1, 0];
+	let bits: &BitSlice<Local, usize> = bits![1, 0, 1, 1, 0, 1, 0, 0, 1, 0];
 	assert_eq!(bits.len(), 10);
 	assert!(bits[0]);
 }